requirements.txt
Flask==2.0.1
tensorflow==2.5.0
requests==2.25.1
PyYAML==5.4.1
